% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_enm.R
\name{extract_enm}
\alias{extract_enm}
\title{Extracts summary statistics from a raster file based on the boundaries of a shapefile}
\usage{
extract_enm(enm, geoshape, id0, id, th, save.plots = FALSE)
}
\arguments{
\item{enm}{input raster data (usually a species distribution model, SDM)}

\item{geoshape}{input shapefile with boundaries of geometries to summarize the raster data for}

\item{id0}{part of \code{'geoshape@data'} that holds the names of the geopolitical units to focus on (polygons of interest)}

\item{id}{subset in \code{geoshape} identifier table to iterate through: a vector of names, may need to match this to a particular part of geoshape with \code{id0}}

\item{th}{threshold of suitability used to calculate RSA. If not provided, the RSA is not calculated.}

\item{save.plots}{save the density plots of points for each \code{id} iterated across (default is FALSE)? If so, the results are stored as a second element in the list \code{[[2]]} or \code{$plots}}
}
\value{
Dataframe (tibble) that contains columns that include
}
\description{
The function \code{extract_enm}
}
